Research + Notes for react setup complete
Researched and took notes on what type of app is best for our project considering the amount of api calls, node.js, and our experience level. Decided on CRA also figured out a file structure for a react project that's simple but also scalable. Researched how to setup the react app in the first place and more.