SCRUM-84 Researched + Notes for React state management complete
Researched, took notes, and determined the different states data can be handled and shared for our app. Then applied the concepts to the features of our app and got a grasp on the use of the frameworks on specific elements of our app like buttons, user specifics, and api data.